… src/ImageSharp/Processing/Processors/Transforms/Resize/ResizeKernelMap.cs
… tests/ImageSharp.Tests/Processing/Processors/Transforms/ResizeKernelMapTests.ReferenceKernelMap.cs