@saucecontrol 2024-08-21
Brain's not fully awake yet today, but I'll give the maths a look soon.