@saucecontrol 2024-08-23
It looks to me like the only differences are due to the change to single precision for kernel normalization and for calculation of the distances passed to the interpolation function.  You'll definitely give up some accuracy there, and I'm not sure it's worth it since the kernels only have to be built once per resize.  You can see [here](https://github.com/SixLabors/ImageSharp/commit/b745ffaf5e2c7c3529f0c82573ed899fcdbeab5b) that @antonfirsov changed the precision to double from the initial implementation some years back.

Since the periodic kernel map relies on each repetition of the kernel weights being *exact*, I can see how precision loss might lead to some differences when compared with a separate calculation per interval.  I've actually never looked at your implementation of the kernel map before, and now my curiosity is piqued because I arrived at something similar myself, but my implementation calculates each kernel window separately, and only replaces that interval with the periodic version if they match exactly.  Part of this was due to a lack of confidence in the maths on my part, as I only discovered the periodicity of the kernel weights by observation and kind of intuited my way to a solution.

@antonfirsov would you mind filling in some gaps on the theory behind your periodic kernel map implementation?  Did you use some paper or other implementation as inspiration, or did you arrive at it observationally like I did?