@JimBobSquarePants 2025-06-12
> > @antonfirsov if you do have any insight I'd appreciate
> 
> Not without going deep down the rabbit hole :(

I thought that might be the case. I'll leave this hanging around for a bit longer, but I don't know if it's worth it. I can do a few smaller things instead (like vectorize normalize.)